requirements.txt
smbpasswd
python-ldap==2.3.13
django
django-annoying
git+https://laqie@github.com/laqie/django-ldapdb.git
django-crispy-forms